Fixes: Allow the intervalPeriod to be changed, for some reason the spinbox had been permanently disabled.
Also removed the logic that disabled interval checking when push email was enabled (and vice versa). It has become apparent that IMAP IDLE has a limitation in that it only checks one folder per connection (e.g. the INBOX), so unfortunately there is still a use case for using interval checking and IMAP IDLE together.

This will change if we (and IMAP server developers) implement support for IMAP NOTIFY, the successor to IDLE.