Description: Reimplement interval checking for IMAP accounts.
Generalize idle queuing code so that it can also be used for interval checking to queue mail checks.

When interval checking fires retrieveMessageList will be called with a minimum of 1 (like IDLE).

IIRC mail services is designed in such a way that interval checking must be reimplemented for each account. I guess the idea is for IMAP we use retrieveMessageList() to check mail and for POP synchronize() to check mail. (Still wonder why we can't implement retrieveMessageList with POP, and least so long as UIDL is supported which we used to, and maybe still do require).

Previously I used a gap between interval mail checking so if mail checking took two minutes to finish and the interval was one minute then there would be a gap (pause) between mail checking. Haven't done that this time, will wait to see if anyone complains (if it is a real problem).

There is a possible problem in that if many accounts are configured with the same timeout (e.g. 10 accounts all set to 10 minutes) then a lot of load will be placed on the server simultaneously.

I notice there is still a roaming enabled checkbox in the .ui file. But I can't use it because of the lack of a mobility services API.