Description: Optimize update exporting so we don't do a full listing (uid search) of all messages in a folder.
Instead just list messages on the server that are going to be operated on (deleted or marked as read). Really we don't need to list any messages, but that would require ignoring errors for missing messages. I'm ok with ignoring errors for uid store commands, but not for select commands (too risky). So I've gone with this partial listing approach.