Several IMAP IDLE (push email) fixes
* Fix a problem with trying to select a non existent folder.
Elaborating: sendSelect now takes a QMailFolder, the first time a connection is made to a push enabled account no folders will exist in the mail store, so we can't expect QMailFolder for the INBOX for this account to exist.

* Check for a non responsive server
Elaborating: When we send a DONE command, make sure we get a response. This should also pick up on a server disconnecting us. Historically we don't appear to have detected disconnects (e.g. on mail.nokia.com). Maybe we need to connect to the socket disconnect signal to catch this immediately.
Also should now detect if an idle connection is established, but we never get as far as issuing a IDLE command (because LOGIN or CAPABILITY command hung).

* Eliminate a race condition where new mail could arrive undetected
Elaborating: After an IDLE connection is disconnected, and then re-establish, we need to start a new mail check i.e. emit idleChangeNotification.