Fix retrieval performance regression.
Retrieval performance decreased with respect to the number of files
in the content directory, due to the use of QDir::entryList to find
files related to the old content identifier when a message was
updated.  Part files are now stored in a per-message subdirectory
of the content location.

The Qtopiamailfile content manager performs an automatic migration
from the old format to the new one - to support this, the
QMailContentManager and QMailContentManagerFactory classes have
an init() function which is called by the mail store after it
completes its own initialisation.