We are only interested in EXISTS idle responses, ignore other responses.
Other responses I've observed are RECENT, which accompanies EXISTS, and provides no additional useful information from our viewpoint. And FETCH which provides info on flag changes (on fastmail but not supported by our cyrus test server), but sadly this response only provides a sequence number not a uid. So we can't make use of it either, and as idle synchronization uses retrieveMessageList (rather than synchronize) we won't pick up flags changes anyway, so we might as well ignore it, rather than trigger an unnecessary mailcheck.

We could possibly pick up flag changes if both QMF and the server supported CONDSTORE.