Refresh flags of messages when list of messages is scrolled, or modified. But not more than once every maximumRefreshRate seconds. And when refreshing flags don't just get flags for visible messages, also get flags for a bufferWidth items around visible messages, so we don't have to retrieve flags for messages when scrolling within the buffer zone.
Prevent retrieval of visible message flags when retrieval is already in progress, as it was causing a deadlock. Also prevent retrieveMoreMessages when retrieval is already in progress, for a similar reason. Ideally requests for retrieval should be queued and compacted, that's an item on the low priority todo list.

Perhaps, we should poll for flag changes even if the list isn't scrolled or modified, maybe as part of interval checking.