Imap IDLE (Push Email) related changes
Don't refresh all folders when new mail is detected instead only refresh the folder the new mail was detected in.
When a new idle connection is established on a folder, refresh that folder.
After modifying an account and enabling push email refresh all folders.

'refresh' a folder means call retrieveMessageList on that folder with a minimum of 1.

Also fix a bug whereby RetrieveMessageList wasn't listing folders with descending=true when a retrievingAllFolder i.e. !QMailFolderId.isValid(). The folder listing during get all mail in the client is probably unnecessary now.
Also fix a race condition in ImapService::Source::setStrategy whereby if a strategy immediately completed then setStrategy would set unavailable to true after retrievalCompleted had already been called.
Also disable the check in ImapRetrieveMessageListStrategy for whether uidNext and exists are the same, for the case when minimum equals one, because it's a problem for push email, due to the don't select an already selected folder optimization. I should maybe issuing a STATUS command here.