Reduce memory used by retrieveMessageList to collect mail, in the case of collecting mail for a folder where the mail store already has many messages stored for that folder.
Preliminary tests show less than 50 bytes of 'useful heap'
per message in the store for the folder being checked.

Details:

Use QMailStore::countMessages instead of loading a list of
messages into memory and counting that list.

Add new minimum/maximum methods to IntegerRegion.

In retrieveMessageList create and maintain a map of min/max
server uids for each imap folder. Make it permanent using custom
folder fields.

Make sure the min/max map is only updated after mail has
been successfully retrieved, (we don't want the IMAP code
to act as if messages have been retrieved that haven't as
this may cause mails not to be retrieved when they should
be. But the opposite case will only cause mails to be
retrieved more than one).

Use the the min/max server uid values to determine which mails
on the server have already been retrieved, previously an in
memory list of all server uids in the folder in the mail
store was made. This requires that only a contiguous list of
messages is keep for each folder in the mail store (no gaps),
which should currently be the case.