Fixes: Mailstore failures when using synchronize or retrieveAll on large accounts.
Description:
Work around limitation in mailstore of not being able to handle more
than 1000 uids in QMailMessageKey::serverUids.

Unfortunately the synchronize and retrieveAll strategies will
again use an increasing amount of memory as the number of mails
in the mail store grows.

retrieveMessageList and exportUpdates should still be ok.

updateMessagesFlags should be ok, as long as less than 250
messages are queried. Which should cover existing use cases.

Ideally the limitation in the mailstore should be eliminated.

I've also removed the workaround to detect messages reappearing on
exchange server. I think the actual problem is when a message is read
with outlook the message is actually deleted and assigned a new imap
uid.