Revert "Fixes: Mailstore failures when using synchronize or retrieveAll"
This reverts commit fab4ab8b3f39d90624c0b887cf856a085672e922.

Restore the use of mailstore comparisons when finding relationships
beteen UID sets, now that the mail store uses temporary tables to
compare sets that are too large for bound parameters.

Removing the 'removed' flag from messages that have previously been
marked is not restored, as this is not believed to be useful.