Prevent the connection from being abruptly closed during push email operations.
In ImapClient::retrieveOperationCompleted() call deactivateConnection()
which starts the logout timer, before emitting signals like retrievalCompleted()
(rather than after).

If the deactiveConnection() is called afterwards, then the connection
is closed during a retrievalMessageList operation that was started
by the imapservice via a connection to retrievalCompleted() that was
triggered as a result of a (recently added) retrieveFolderList i
operation completing.

May also be useful for QMF clients that do something similar.