Recover from severed connections.
If a connection is lost while executing an imap command then the state of the fsm used by the imap protocol will be left in the last active state e.g. select.

So that when an attempt to reconnect is made by ImapProtocol::open, the _fsm->setState(&_fsm->initState) statement will fail with the 'Unable to issue command simultaneously:' error as _fsm is still in the last active state e.g. select.

So in ImapProtocol::open call _fsm->reset() before setting the state to initState.

I could have called reset when the severing of the connection is discovered but I'm not sure I can catch all the cases, e.g. when the fsm is in the logout state.