Do not exclude internal header fields with Identity format.
Also, if setBody() is used to add a body to a message where the
content type indicates a multipart message, the actual body element
will be ignored.  Thus, the actual multipart variant is parsed
by the MIME code, and the result is equivalent to having set the
metadata elements explicitly.