Fix bug in get more messages functionality.
There is a customer field qtmail-retrieved-minimum which stores the
number of messages already retrieved for a folder.

This qtmail-retrieved-minimum optimization results in incorrect results
in the case when qtmail-retrieved-minimum is computed for a folder and
then new messages are retrieved in that folder.

(As the qtmail-retrieved-minimum value is not updated to take into
consideration the newly added messages).