Fixes:    Push email not collecting all mail
Task:     QTSOL-85<task number>
Details:
More testing is required. Working on a couple theories:

1) There's a failure to add a message to the store. But this error is
ignored and the qmf-max-serveruid folder custom field is updated
anyway.

2) Stale _newMinMaxMap values are being used.

Added a bool _error to ImapStrategy, so that
qmf-max-serveruid is not updated if an error occurs during a
store operation while the retrieveMessageList strategy is executed.

Made sure to clear _newMinMaxMap in ImapRetrieveMessageList::LogIn.

It would be nice to know why I'm getting the fail on store operations