Fix crash in servicehandler.cpp when dereferencing a deleted pointer
Patch contributed by Mills Matthew. "If a QMailMessageService instance
has been freed but that service still is associated with an action then
we’ll have trouble when that action expires. Specifically, this
trouble occurs in ServiceHandler::expireAction when we are iterating
through and acting on the ActionData::services of an expired action."