Fixes:    Retrieving an attachment causes full mail check,
Task:     QTSOL-154
Details:  In the EmaiClient class the retrievalCompleted slot checks the mailAccountId member variable, and if it's valid assumes that a mail check is in progress and advances from the exportUpdates to retrieveFolderlist stage of full account mail checking. So set mailAccountId to an invalid id when retrieving attachments and performing other types of non full account mail checking retrieval.

I got an unable to allocate qmailserviceaction error in qtmail before making this change. Not sure if it's difficult to reproduce or if it's fixed with this change.