Fix add/updateMessages list handling.
There are some lists that are actually "global" (for the
x add/update operations) sets. The code was not treating
the lists in this way at all. Some places assumed the
lists were cleared for each item. Some places assumed
it was ok to append without checking (leads to
duplicates).  Some places assumed it was ok to overwrite
the list (instead of appending to it).

This patch should address all of these concerns.

What is missing though is an addition to the unit test
that demonstrates each of the places where a bad
assumption was made. I am primarily relying on visual
inspection.

Fixes: QTSOL-166