qMailLog can use a runtime check.
To enable a category, ~/.config/Nokia/QMF.conf must be created.
Most categories can be enabled by creating the file with the
following contents:

[General]
Messaging=1
IMAP=1
SMTP=1
POP=1

The messageserver process will read this file the first time it
encounters a qMailLog() call. To force it to re-read the file
you must send it the HUP signal.

Note that to enable the ImapData or MessagingState categories you
must edit qmaillog.h or build with QMF_ENABLE_LOGGING.

Fixes: Maemo Bugzilla 176463