Move disconnected move/copy code out of qmailserviceaction.cpp and back into imapservice.cpp. (This is a no-op for clients that do not perform disconnected moves/copies).
There's a not so obvious problem with having them in
qmailserviceaction.cpp, if a client performs multiple
operations quickly, and both these operations require
'flushing' pending disconnected operations to the server,
then the later operation (service action) will try to
operate (e.g. move) messages that are being dealt with
by the first action. After the first action is complete
and the second action begins to execute, the second action
will fail because it can't move messages that have already
been moved by the first action.

So the solution is to not generate the list of messages to
act on until as late as possible, namely when the action
beings executing.