Tweak runtime logging in debug builds.
Previously doing a debug build would turn the runtime-enabled
logging categories on unconditionally. This is changed so that
the _default_ for logging categories is on in debug builds. This
means that QMF.conf can now turn off logging categories in debug
builds.

Release builds see no change in behaviour. Debug builds will see
a change only if they have a QMF.conf that turns off categories.