Replace use of QMessage:setId and updateMessage, with removeMessage and addMessage.
Fixes a problem with message files in the $QMF_DATA/tmp directory being deleted randomly, but very rarely.

A workaround is already in place to address this issue, so no actual behavior change should occur, but this change addresses the root cause.

If retrieval of messages was terminated during retrieveMessageList during previewing due to a network time out or user cancelling etc, then preview message files would remain. A subsequent successful mail check would replace each left over preview file with a new message by creating a new message using fromRfc2822, setting the messageId, and calling updateMessage. These new messages unfortunately wouldn't have a contentScheme specified so updateMessage wouldn't use the contentmanager to update the message file, instead only the metadata would be updated. The metadata would contain a detached-filename custom field which is normally stripped out by the contentmanager but in this case the contentmanager would not be called (due to a missing contentScheme) so the detached-filename field would not be stripped out and instead would be saved to the database. Later, after the previewing phase of message retrieval during message completion the preview would be loaded and updated, the existing detached-filename would be loaded and used to rename a file by the contentManager, but as the detached-filename was stale, and as it refers to a random file name generated by QTemporaryFile any file in the tmp directory that coincidentally had the same name as the stale detached-filename would be renamed. This could cause (very rarely) the buffer for a message being retrieved to be renamed during message retrieval.

Arguably the *::setId functions should be private, to eliminate the possibility of a plugin setting the messageId but not the contentScheme.