Extensions to the logging subsystem (From Vitaly Repin)
Taken from QMF merge request #6 by Vitaly Repin:

Current version of the QMF logging subsystem has the following issues:

1) Whenever application is linked against QMF, all the syslog messages are
prefixed with QMF regardless of the application settings.

E.g., Aug 31 00:25:16 vitnote QMF[4959]: [Critical] Test

It makes it harder to identify the application which has printed specific
message to syslog. It also overwrites the previous calls to "openlog" made by
application which is impolite

2) It is not possible to enable/disable logging through syslog in runtime
3) It is not possible to log into file and stderr
4) Output of Q_ASSERT is not logged. In fact, QWarning(), qFatal(), qDebug()
do not go to the log
5) Library installs the SIGHUP handler transparently for the application. It
is bad because application might want to install its own handler for this
signal and it could conflict with the library's logic.

The patch attached for your review fixes the issues above. It also adds the
following possibilities:

- To log into file and stderr, controllable through the configuration file in
runtime
- To create chain of loggers if needed (e.g., output to file and stderr, to
file and syslog etc). Chain of loggers is defined in run-time through
configuration file.
- Configuration file is per application. Not the only file for all the users
of the library.