IMAP message deletion speed up
The basic idea is to detect duplicates by getting the serverUid of each
message. Then in one hit, I can check if there are duplicates. If none,
proceed as normal.

If there are duplicates, I now have a list of a ids and serverUids -- so
I can [slowly] iterate through through the lists of ids and find the
offenders. It should be significantly faster (no list copy, 2 (albeit
more complex) queries vs 1+ids.size()), and negligibly faster in the
case of duplicates (only need to retrieve serverUid rather than
serverUid and Id)