Fix, NB#203640, add 'pushCapable' and 'capabilities' fields to imap4 service configuration.
These fields are updated when an imap account is logged into on an external server.

pushCapable will be set to "0" if the imap server doesn't support the IDLE capability,
otherwise pushCapable will be != "0".

capabilities is set to a space separated list of capabilities as returned by the IMAP CAPABILITY command.

As an example the pushCapable field can be read as follows for an account with id accountId:
  QMailAccountConfiguration config(accountId);
  QMailServiceConfiguration serviceConfig(config, "imap4");
  bool pushCapable(serviceConfig.value("pushCapable", "0").toInt() != 0);