Fix bug with partial message downloading, NB#239506.
From Gabor.Fekete@tieto.com

"With regards to partially downloaded parts, previously QMF continued
a partial download ONLY when a "minimum" amount of bytes were requested.
When "minimum" was set to "SectionProperties::All", the start and end positions
were not initialized, so the whole content was downloaded and this "new" content
was then appended to the previous existing partial content. This resulted in
duplicated content in the beginning of a part."

From