Possible fix/workaround for NB#253783
Remove 'static const' from 'static const QRegExp' expressions.

Because 'Static non-PODs aren't a good idea'
 http://old.nabble.com/Multiple-QRegExp-crashes-when-multithreading-in-KRunner-td30666724.html

(As order of static initializers is not well defined?)

As Qt 'maintains a list of the most recently constructed QRegExp objects to
avoid the cost of recompiling the same regular expression', the 'static const'
should probably be superfluous from a performance POV.

'static const' is not thread safe, but QMF isn't anyway so this is a
minor/irrevelant point.