Better error handling if messageserver fails.
If messageserver fails all running actions will immediately fail.

If an action is started while the messageserver is not running, then
after a timeout period that action will fail. However if messageserver
starts before the timeout period expires then the action will be
dispatched to the messageserver.

Related to NB#218816