On any place in qmf where we only accepted CRLF endlines, now we also accept LF.
This fix is for some bad servers (like sohu.com IMAP) that don't honor standard
and use LF as end of line instead of CRLF. This was causing huge problems
parsing the responses from server (i.e. on retrieving the body of a message
we would always get the headers rewritten with empty fields, and also the body
parts wouldn't be properly detected).