Fix NB#276243 Email Sync fails when not synced for a day...
Fix: Email Sync fails when not synced for a day and shows connecting to
server always. Restarting device resolves the problem temporarily

From Morenkov.

IMAP service disables and then enables itself. This causes
initClientConnections() to be called in which _strategyExpiryTimer is connected
to the slot which shows the text mentioned above. No disconnects are found.

After this, there are 2 connects to _strategyExpiryTimer and its next
expiration causes double call of expireStrategy() slot. And so on... the next
expiration will cause 4 calls... etc. So, IMAP plugin resets itself
continuously.