Fix for NB#271244 inline .txt breaks preview
Patch from Edit:
"The reason why the preview is showing the content of the text attachment is
that the text attachment's mail part does not have Content-Disposition type
Attachment but instead type Inline.
In the QMailMessage::refreshPreview() method it is recognized as a
plainTextContainer() and it's body is taken as preview.

A solution to this could be that in case of Multipart/Related message we look
first for the htmlContainer instead of  plainTextContainer and accepting also
empty body (otherwise in case of no body of the rich text message the text
attachment's body is shown)."