Be more restrictive when enabling QMailAccount::CanTransmitViaReference.
Require CATENATE cpability in addition to existing URLAUTH requirement.

Because when FWOD'ing if a sent folder on server exists then CATENATE
will be used to construct a message in the sent folder.

Not actually aware of any server that supports URLAUTH but not
CATENATE so this is just being safe.