Upgrade IMAP accounts to set QMailAccount::CanSearchOnServer flag
When an IMAP service is instantiated, e.g. when sync is done in
an IMAP account.

Normally an upgrade script like mailaccounts-106-107.sqlite.sql
would be used to do this, but it would be time consuming to do
as the sql would need to set the status bit depending on the
type of the account, including MfE accounts. (Not sure what
account type string MfE accounts have).