Adjust the pending-data check for ssl sockets.
Patch form Gabor Fekete.

In QSslSocket::encryptedBytesWritten(qint64 written), "written" may be less (or
more) than the amount of the original (unencrypted) data.
Unfortunately, it seems that QSslSocket::encryptedBytesToWrite() returns 0
both right after writing to the socket (via QDataStream) and also during
the QSslSocket::bytesWritten(qint64) signal. This makes it quite impossible
to do a reliable bookkeeping of sent data, as is done with QTcpSocket.

So, we resort to watching the QSslSocket::encryptedBytesToWrite() and
QSslSocket::bytesToWrite() values to detect when all of the data SHOULD
have been sent.