Fix for NB#295637. Improve metadata headerfield decoding.
Decode the metadata headerfield using the charset of the plain text or
html body part if one is found, and the libicu analysis fails to find a
 charset with a high confidence, and the mail doesn't contain a charset
parameter in the content type headerfield.

The plain text or html body part may not be found in some situations,
e.g. depending on user settings, only headers may be downloaded, and
for IMAP only IMAPSTRUCTURE envelope may be downloaded, or two passes
may occur and only envelope may be downloaded in first past.