BIC: Integrate charset fixes from Harmattan
New function static QByteArray QMailCodec::bestCompatibleCharset(const QByteArray& charset, bool translateAscii);

commit 227635e77381bf98158b6d373a40760877d47e9b
Author: Javier Fernandez <jfernandez@igalia.com>
Date:   Fri Nov 4 11:20:19 2011 +0100

    Fixes: NB#289443 - Mail; Hong Kong special characters are displayed as garbage text.

commit d7bce4800db29006c5e0da1b5a34598a5f44f911
Author: Javier Fernandez <jfernandez@igalia.com>
Date:   Fri Oct 28 11:06:10 2011 +0200

    Fixes: NB#286730 - Thai texts cannot be read in gmail.

NB#294947 - Mail; Preview in Hong Kong special Chinese characters (Big5) displays as garbage after downloading the attachment.

Apply the bestCompatibleCharset function to the ensureCharsetExists

when the preview is generated it checks if the body is encoded or not. If not
encoded, it does not call to the codecForName, which actually replaces the
charset (GB2312 --> GB18030) to one more appropriated or general.

In that case, the charset coming into the body is used, so I think it's a good
idea to do it inside the ensureCharsetExists, as it's called every time the
body is built in memory.

We could just solve the problem by calling the new method inside the
QMailMessageBodyPrivate::toStream(QTextStream& out) const method, used to
generate the preview form a plain text body. However, there are additional
places where the charset is used without checking for a proper replacement, so
calling it inside the ensureCharsetExists method seems a better approach.