BIC: Fix a bug in data associated with QMailMessageMetaData::recipients
Patch from Edit Siket-Szasz.

QMailMessageMetaData::recipients only contained to recipients,
whereas it should also contain cc and bcc recipients, and should
do so after this commit is made.

This change will appropriately update the data stored in the
recipients row of the mailmessages table in the qmailstore.db
database for new mails, but no retroactive changes will be made.

Hence the behavior of the QMailMessageKey::recipients functions
will be modified, and should now be correct.

The following API changes have been made as documented in the
qmfclient library CHANGES file.

49.  Added QList<QMailAddress> QMailMessage::to() const
           virtual QList<QMailAddress> QMailMessageMetaData::recipients() const
        protected:
           virtual void QMailMessageMetaData::setRecipients(const QList<QMailAddress>& s);
           virtual void QMailMessageMetaData::setRecipients(const QMailAddress& s);

     Removed QList<QMailAddress> QMailMessageMetaData::to() const
             void QMailMessageMetaData::setTo(const QList<QMailAddress>& s);
             void QMailMessageMetaData::setTo(const QMailAddress& s);

This change is BIC (a binary incompatible change).