Fix qtmail bug in editing message in drafts folder on server
Updated message should now be saved to drafts folder on server.

This required deleting old message, and saving a new message. Which
has some side effects. #1 On gmail deleted messages are copied to the
trash automatically. #2 In qtmail if multiple qmf accounts exist
for the same remote account then updating a drafts in one qmf
account and then viewing the other account will detect that the
original drafts message has been removed due to qtmail automatic
visible message flags updating, and the new message
will not be retrieved until a synchronization is performed.