Better prevention and handling of loops in message conversations.
In messagePredecessor don't allow loops in conversations.

In identifyAncestors and resolveMissingMessages don't allow a
trivial conversation loop of a message being it's own parent.

Fix logic bugs in QMailStorePrivate deleteMessages, deleteFolders,
and deleteAccounts. These functions were expecting empty 'updated*'
parameters, but this was not the case. e.g. deleteAccounts,
calls deleteFolders, which updates the updatedMessages list, and
then deleteAccounts calls deleteMessages with the non-empty
updateMessages list. This was triggering an asserts or in release
mode causing an infinite loopi (maybe only in the case of a
conversation loop existing).

In QMailMessageThreadedModel don't go into an infinite loop if
there is a conversation loop.