Reduce frequency of attempts to establish push email connections.
Was previously trying about once a minute, which causes high battery
drain. Now try with a 0 second delay, then a 30 second, delay then
double up to a maximum of a 1 hour delay.

A push connection may not be able to be established if the remote
server goes down, or there are connection problems, can be exacerbated
by password changes.