FWOD fix, don't try to upload email twice to remote server
When sending a forward without download message, with a sent
folder on the server, a second spurious attempt was being made
to APPEND the message into the sent folder on the server. (Test
server ignored the spurious append, so no error was detected).

So fix this, checked for regressions in 4 cases:
 Forward without download with sent folder on server.
 Forward without download without sent folder on server.
 Normal sending (non-fwod) with sent folder on server.
 Normal sending without sent folder on server.