Fix bug with pop mails not being deleted from pop server.
When syncing pop service was not waiting for response from
server after sending a QUIT command.

Fix by updating pop service logic as follows, if mails are
to be deleted from server then issue a QUIT command immediately
after deleting mails. This is because after issuing a DELE
command a QUIT command must be issued to force the POP server
to really delete mails.

Otherwise if not mails are deleted then keep the connection
open for several seconds, so that if a client makes another service
request, e.g. to fetch the body of a mail left on the server then
the existing connection can be reused.

However as an exception if a sync (retrieveMessageList) is
requested, when the connection is already open, then the
connection is closed and reopened, because this is necessary
in order to see new mails listed.

To test this change a tried several different uses cases
1) Sync after deleting (empty from trash) mails using device
2) Sync after not deleting mails
3) Do 1 then quickly sync before connection is closed
4) Do 1 then sync after connection is closed
5) Do 2 then quickly do 1 before connection is closed
6) Do 2 then do 1 again after connection is closed
7) Do 1 then download body of message left on server