Add new QMailMessage::NoNotification flag
This flag indicates that a new message notification should not be shown
for the message. e.g. an older message retrieved in a folder that has previously
been synchronized, or an existing message moved to a folder such as trash, or a
message externalized by saving in a drafts or sent folder

It should only be used for new messages, e.g. messageAdded signal.

(I've also included some missing doc in the commit.)