Fix bug in QCopClient::handleDetachMonitor
It was being removed from serverMap which doesn't make
sense as it is an iterator on serverMonitorMap which is a completely
different object.

Additionally inspection of the function logic alone indicates
this is a typo and semantically it should be removed from
serverMonitorMap.

Not actually used by QMF except in the tst_qcop automatic test.
Caused crash in test on arm architecture but not intel.