search QMF_INSTALL_ROOT for plugins
the most likely place for plugins to be is in the place where they were
configured to be installed, so search there.

this avoids the need for all users of QMF to ensure an environment variable is
exported before use.

the old fallback ("." searching) should still work, as:
a) we never explicitly passed "." anyway
b) QCoreApplication's library paths internally contain the current directory of
   the application

Change-Id: I50d70672b11fc736c9d65d95e200d197ae119394
Reviewed-by: Valerio Valerio <vdv100@gmail.com>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>