Remove generated HTML documentation.
The original reasoning for including this in the repository in
bdf5d879572e5a835bb5e26c6c100680899ad67c was that the README referenced it for
install/run instructions. Rectify that by moving the relevant documentation to
README.

This is a lot more optimal than having a 3+mb snapshot of HTML documentation
which requires periodic updating.

Change-Id: I90a414da2019b035ae99c8aa6ea4b1ceeb41eb6e
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>