Convert qmfclient & qmfmessageserver into Qt modules: QmfClient & QmfMessageServer.
This was not possible prior to now due to Qt 4 support hanging around like the
undead vampire it is, but now we purged it with fire and light, and the build
system looks a lot less like a stinky, decomposed corpse.

There's still a lot of surprises lurking: qtmail is one such surprise. I don't
quite grasp why it's so heavy on plugins, and qmfutil also needs some...
attention. But at least it's a start.

As part of this work, we rename the qcop headers to fit the private convention.
This could have been done in a prior commit, but given that I didn't have an
easy way to build prior to this commit, I opted to roll them together.

Change-Id: Ia3e288ffc3639a7751c9040ceecb54fca77a31b1
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>
Reviewed-by: Valerio Valerio <valerio.valerio@jollamobile.com>