Remove composer & viewer plugins, move them entirely to QtMail.
These may have made some sense back when QMF was part of Qtopia (a messaging system
customizable for each platform and offering extensible message types), but it
makes a lot less sense in the sense of a library offering email-related
functionality.

This type of functionality belongs in the clients, not in the library, so move
it there.

This has been performed through a fairly quick hatchet job. It isn't really
clean, working code yet, but it is operational.

Change-Id: Ic417a5e885034b67b645df238febb30cda88b17c
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>
Reviewed-by: Valerio Valerio <valerio.valerio@jollamobile.com>