Replace QT_NO_OPENSSL with QT_NO_SSL.
This macro was introduced in Qt 5 development to allow for non-OpenSSL backends,
so the presence of OpenSSL doesn't now necessarily reflect the availability of
SSL.

Some of this looks like it could use some further attention (e.g. it seems a bit
nasty to modify public API based on the presence of SSL where that can be
avoided) but that is work for another day.

Change-Id: Ia9006e43aeb1755f17ebdafb629c11f924263c3d
Reviewed-by: Richard J. Moore <rich@kde.org>
Reviewed-by: Valerio Valerio <valerio.valerio@jollamobile.com>