Reset stream status when writing message from memory to file
QDataStream::readRawData() is setting the status of the stream to
ReadPastEnd (an error status) even though the return value from the
function is positive (i.e. not an error value) and the stream is
merely at the end after the call, in the case where the len parameter
to the method is larger than the (QByteArray) data backing the stream.

This commit ensures that we detect this case and reset the stream
status (to Ok) so that processing can continue.

See QTBUG-69474 for more information

Change-Id: If2244f663c671046b9bb2b09076f069f50e591a2
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>