Move contentDisposition and contentDescription into QMailMessagePartContainer
According to RFC1806, see 3. Examples, it is valid to set content-disposition
to a message itself, or even to a multipart container. Currently, QMF cannot
represent a message as given in the first example, because QMailMessage inherits
only QMailMessagePartContainer but the content disposition is stored in
QMailMessagePart.

This change would allow to treat a message whose only content is marked as
an attachment.

Change-Id: Ifeada93dfca7e1a351c1a73049e9209d3eacd2fc
Reviewed-by: Michael Nosov <Michael.Nosov@harman.com>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>