Use codec detection when charset is unknown
For emails where the plain text (or the HTML version) is declared with an unknown charset,
the body()->data() of the text part (or the HTML one) is QString() empty.

 Example of wrong declaration:
 --0ec52789c76dd9f8e88d72dc967821ec
 Content-Type: text/plain; charset=quoted-printable
 Content-Transfer-Encoding: quoted-printable

Emails like that are definitely violating the standard and are indeed broken. But having
an empty display when viewing the email is not useful neither.

The issue is in QMailCodec::decode(), where the "if (codecForName())" has no fallback if
the codec does not exist. This commit is adding a fallback, assuming UTF-8 encoding.

Change-Id: I7b67cea999a2c1eb4211749d87a4ccb54daa332d
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>